+++ packages/blade/src/components/Input/DropdownInputTriggers/SelectInput.stories.tsx
@@ -2,18 +2,20 @@
 import type { StoryFn } from '@storybook/react';
 import { Title } from '@storybook/addon-docs';
 import React from 'react';
-import { SelectInput } from './SelectInput';
+import { bladeTheme } from '~tokens/theme';
+import { BladeProvider } from '~components/BladeProvider';
+import { Badge } from '~components/Badge';
+import { BankIcon, InfoIcon } from '~components/Icons';
 import iconMap from '~components/Icons/iconMap';
+import { SelectInput } from './SelectInput';
 import { Sandbox } from '~utils/storybook/Sandbox';
 import StoryPageWrapper from '~utils/storybook/StoryPageWrapper';
 import { Dropdown, DropdownOverlay } from '~components/Dropdown';
 import { ActionList, ActionListItem, ActionListItemBadge } from '~components/ActionList';
 import { Box } from '~components/Box';
-import { Badge } from '~components/Badge';
 import { Text } from '~components/Typography';
 import { Link } from '~components/Link';
 import { Tooltip, TooltipInteractiveWrapper } from '~components/Tooltip';
-import { BankIcon, InfoIcon } from '~components/Icons';
 
 const propsCategory = {
   BASE_PROPS: 'Select Input Props',
@@ -234,7 +236,7 @@ const SelectInputTemplate: StoryFn<typeof SelectInput> = ({ icon, ...args }) => 
           onChange={({ name, values }) => {
             console.log(name, values);
           }}
-          icon={iconMap[(icon as unknown) as string]}
+          icon={iconMap[icon as unknown as string]}
           valueSuffix={({ values }) => {
             if (values[0] === 'item-1') {
               return <Badge color="positive">20% Off</Badge>;
@@ -254,6 +256,15 @@ const SelectInputTemplate: StoryFn<typeof SelectInput> = ({ icon, ...args }) => 
           </ActionList>
         </DropdownOverlay>
       </Dropdown>
+      <BladeProvider themeTokens={bladeTheme} colorScheme="dark">
+        <Box
+          marginTop="spacing.11"
+          marginLeft="-12px"
+          backgroundColor="surface.background.gray.intense"
+          height="100px"
+          width="100px"
+        />
+      </BladeProvider>
     </Box>
   );
 };