+++ packages/blade/src/components/ListView/docs/ListView.stories.tsx
@@ -543,7 +543,6 @@ const Page = (): React.ReactElement => {
                         <TableFooterCell>-</TableFooterCell>
                         <TableFooterCell>-</TableFooterCell>
                         <TableFooterCell>-</TableFooterCell>
-                        <TableFooterCell>-</TableFooterCell>
                         <TableFooterCell>
                           <Amount value={10} />
                         </TableFooterCell>
@@ -982,7 +981,6 @@ const DefaultExample: StoryFn<typeof ListView> = (args) => {
                   <TableFooterCell>-</TableFooterCell>
                   <TableFooterCell>-</TableFooterCell>
                   <TableFooterCell>-</TableFooterCell>
-                  <TableFooterCell>-</TableFooterCell>
                   <TableFooterCell>
                     <Amount value={10} />
                   </TableFooterCell>
@@ -1315,7 +1313,6 @@ const ControlledExample: StoryFn<typeof ListView> = (args) => {
                   <TableFooterCell>-</TableFooterCell>
                   <TableFooterCell>-</TableFooterCell>
                   <TableFooterCell>-</TableFooterCell>
-                  <TableFooterCell>-</TableFooterCell>
                   <TableFooterCell>
                     <Amount value={10} />
                   </TableFooterCell>
@@ -1397,7 +1394,11 @@ const WithBulkActionExample: StoryFn<typeof ListView> = (args) => {
   const isMobile = useIsMobile();
 
   return (
-    <BaseBox height="100%">
+    <BaseBox 
+      backgroundColor="surface.background.gray.moderate"
+      padding="spacing.8"
+      height="100vh"
+    >
       {isMobile && (
         <SearchInput
           label=""
@@ -1800,7 +1801,11 @@ const MultiSelectQuickFilter: StoryFn<typeof ListView> = (args) => {
   const isMobile = useIsMobile();
 
   return (
-    <BaseBox height="100%">
+    <BaseBox 
+      backgroundColor="surface.background.gray.moderate"
+      padding="spacing.8" 
+      height="100vh"
+    >
       {isMobile && (
         <SearchInput
           label=""
@@ -2081,7 +2086,6 @@ const MultiSelectQuickFilter: StoryFn<typeof ListView> = (args) => {
                   <TableFooterCell>-</TableFooterCell>
                   <TableFooterCell>-</TableFooterCell>
                   <TableFooterCell>-</TableFooterCell>
-                  <TableFooterCell>-</TableFooterCell>
                   <TableFooterCell>
                     <Amount value={10} />
                   </TableFooterCell>
@@ -2355,7 +2359,6 @@ const WithoutSearchExample: StoryFn<typeof ListView> = (args) => {
               <TableFooter>
                 <TableFooterRow>
                   <TableFooterCell>Total</TableFooterCell>
-                  <TableFooterCell>-</TableFooterCell>
                   <TableFooterCell>-</TableFooterCell>
                   <TableFooterCell>-</TableFooterCell>
                   <TableFooterCell>-</TableFooterCell>
@@ -2706,7 +2709,6 @@ const WithDropDownSearchExample: StoryFn<typeof ListView> = (args) => {
               <TableFooter>
                 <TableFooterRow>
                   <TableFooterCell>Total</TableFooterCell>
-                  <TableFooterCell>-</TableFooterCell>
                   <TableFooterCell>-</TableFooterCell>
                   <TableFooterCell>-</TableFooterCell>
                   <TableFooterCell>-</TableFooterCell>