+++ packages/blade/src/components/Popover/componentIds.ts
@@ -1,0 +1,5 @@
+const componentIds = {
+  Popover: 'Popover',
+};
+
+export { componentIds };